+++ components/components.ts
@@ -27,6 +27,7 @@ export type { CascaderOption, CascaderProps } from "./cascader/types";
 
 export { default as Card, CardMeta } from "./card";
 export type { CardMetaProps, CardProps } from "./card";
+
 export { default as FeatureCard } from "./feature-card";
 export type { FeatureCardProps } from "./feature-card";